refactor: move statics out of ISAAC class
